requirements.txt
requests==2.31.0
arxiv==2.1.0
pyzotero==1.5.18
pydantic==2.6.1
openai>=1.30.0
schedule==1.2.1
python-dotenv==1.0.1
streamlit==1.31.0
networkx==3.2.1
pyvis==0.3.2
matplotlib==3.8.2
PyYAML==6.0.1
google-generativeai>=0.5.0
tenacity==8.2.3
importlib-metadata>=6.0.0